+++ packages/ra-core/package.json
@@ -66,7 +66,7 @@
     },
     "dependencies": {
         "classnames": "~2.2.5",
-        "date-fns": "^2.21.1",
+        "date-fns": "^1.29.0",
         "eventemitter3": "^3.0.0",
         "inflection": "~1.12.0",
         "lodash": "~4.17.5",
