+++ packages/ra-no-code/package.json
@@ -46,7 +46,7 @@
     },
     "dependencies": {
         "classnames": "~2.2.5",
-        "date-fns": "^2.21.1",
+        "date-fns": "^1.29.0",
         "inflection": "~1.12.0",
         "lodash": "~4.17.5",
         "react-admin": "^3.14.5",
