+++ packages/ra-no-code/src/ResourceConfiguration/FieldConfiguration/index.ts
@@ -1,0 +1,2 @@
+export * from './FieldTypeInput';
+export * from './FieldViewsInput';
